Abtracted styling jss classnames into the styles.js file in the utils directory
