+++ frontend/src/pages/_sandbox/default/index.js
@@ -4,7 +4,7 @@ import PageWrapper from 'components/layouts/PageWrapper'
 import GlobalNavBar from 'components/navbars/GlobalNavBar'
 import Footer from 'components/layouts/Footer'
 
-import { makeStyles } from '@material-ui/core/styles'
+import useStyles from 'utils/styles'
 import {
     Card,
     CardActions,
@@ -19,52 +19,6 @@ import Container from 'components/generic/Container'
 import Button from 'components/generic/Button'
 import Tag from 'components/generic/Tag'
 import { Skeleton } from '@material-ui/lab'
-// import theme from 'material-ui-theme'
-import theme from 'junctionTheme'
-
-// import { useMyEvents } from 'graphql/queries/events'
-
-const useStyles = makeStyles({
-    w450: {
-        maxWidth: '450px',
-    },
-    shadow1: {
-        boxShadow: '0 2px 4px 0 rgba(0, 0, 0, 0.16)',
-    },
-    border1: {
-        borderRadius: '8px',
-    },
-    borderButton: {
-        borderRadius: '30px',
-    },
-    textAlignStart: {
-        textAlign: 'start',
-    },
-    pb2: {
-        paddingBottom: theme.spacing(2),
-    },
-    p3: {
-        padding: theme.spacing(3),
-    },
-    pxb3: {
-        padding: theme.spacing(0, 3, 3),
-    },
-    fs5: {
-        fontSize: theme.spacing(5),
-    },
-    fs2half: {
-        fontSize: theme.spacing(2.5),
-    },
-    fs2: {
-        fontSize: theme.spacing(2),
-    },
-    fs3: {
-        fontSize: theme.spacing(3),
-    },
-    cardNote: {
-        fontSize: '18px',
-    },
-})
 
 export default () => {
     // const [events, loading] = useMyEvents()
